requirements.txt
skale.py==2.0.dev18
python-dotenv==0.10.3
